README-edits.md
Currency-Currently is going to be a web browser built on the blockchain. In addition to the web browsing capabilities there will be a place for shopping to spend cryptocurrency. A place for crypto currency educational materials. A place for real time crypto market data, news & updates. A place for social network style chat to discuss whats currently going on in the world of crypto currency.